+++ test/WebApi20.FunctionalTests/FunctionalTest/MultipleWebHostsTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using System.Reflection;
 using System.Runtime.InteropServices;
 using AI;
 using FunctionalTestUtils;
@@ -50,36 +51,34 @@ namespace WebApi20.FunctionalTests20.FunctionalTest
             }
         }
 
-        [Fact(Skip = "We track each request and depednency by each WebHost, issue #621")]
+        [Fact]
         public void TwoWebHostsCreatedInParallel()
         {
             using (var server1 = new InProcessServer(assemblyName, this.output))
             using (var server2 = new InProcessServer(assemblyName, this.output))
             {
                 this.ExecuteRequest(server1.BaseHost + requestPath);
-                var telemetry1 = server1.Listener.ReceiveItems(TestListenerTimeoutInMs);
-                this.DebugTelemetryItems(telemetry1);
-
                 this.ExecuteRequest(server2.BaseHost + requestPath);
+
+                var telemetry1 = server1.Listener.ReceiveItems(TestListenerTimeoutInMs);
                 var telemetry2 = server2.Listener.ReceiveItems(TestListenerTimeoutInMs);
 
+                this.output.WriteLine("~~telemetry1~~");
+                this.DebugTelemetryItems(telemetry1);
+                this.output.WriteLine("~~telemetry2~~");
                 this.DebugTelemetryItems(telemetry2);
-                Assert.Single(telemetry1.Where(t => t is TelemetryItem<RequestData>));
-                Assert.Single(telemetry1.Where(IsServiceDependencyCall));
+
+                Assert.Equal(2, telemetry1.Count(t => t is TelemetryItem<RequestData>));
+                Assert.Equal(2, telemetry1.Count(IsServiceDependencyCall));
                 Assert.DoesNotContain(telemetry1, t => t is TelemetryItem<ExceptionData>);
 
-                var request1 = telemetry1.Single(t => t is TelemetryItem<RequestData>);
+                var request1 = telemetry1.First(t => t is TelemetryItem<RequestData>);
+                var request2 = telemetry1.Last(t => t is TelemetryItem<RequestData>);
                 Assert.Equal("200", ((TelemetryItem<RequestData>)request1).data.baseData.responseCode);
-
-                // Fails here, we track everything twice
-                // it did not happen with the first host because second one has not been really started yet
-                Assert.Single(telemetry2.Where(t => t is TelemetryItem<RequestData>));
-                Assert.Single(telemetry2.Where(IsServiceDependencyCall));
-                Assert.DoesNotContain(telemetry2, t => t is TelemetryItem<ExceptionData>);
-
-                var request2 = telemetry2.Single(t => t is TelemetryItem<RequestData>);
                 Assert.Equal("200", ((TelemetryItem<RequestData>)request2).data.baseData.responseCode);
-            }
+
+                Assert.False(telemetry2.Any());
+           }
         }
 
         [Fact]
@@ -127,7 +126,16 @@ namespace WebApi20.FunctionalTests20.FunctionalTest
                 return;
             }
 
+            // Active config could be used multiple times in the same process before this test
+            // let's reassign it
+
+            TelemetryConfiguration.Active.Dispose();
+            MethodInfo setActive =
+                typeof(TelemetryConfiguration).GetMethod("set_Active", BindingFlags.Static | BindingFlags.NonPublic);
+            setActive.Invoke(null, new object[] { TelemetryConfiguration.CreateDefault() });
+
             var activeConfig = TelemetryConfiguration.Active;
+
             using (var server = new InProcessServer(assemblyName, this.output))
             {
                 this.ExecuteRequest(server.BaseHost + requestPath);
@@ -149,6 +157,7 @@ namespace WebApi20.FunctionalTests20.FunctionalTest
                 Assert.Single(message);
 
                 this.output.WriteLine(((TelemetryItem<MessageData>)message.Single()).data.baseData.message);
+
                 Assert.Equal("some message after web host is disposed", ((TelemetryItem<MessageData>)message.Single()).data.baseData.message);
             }
         }