+++ test/WebApi20.FunctionalTests/FunctionalTest/RequestTelemetryWebApiTests.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Collections.Generic;
     using System.Diagnostics;
-    using System.Linq;
     using System.Text.RegularExpressions;
 
     using FunctionalTestUtils;
@@ -106,8 +105,6 @@
                 return builder.ConfigureServices(services =>
                 {
                     services.AddApplicationInsightsTelemetry();
-                    services.Remove(services.Single(sd =>
-                        sd.ImplementationType == typeof(DependencyTrackingTelemetryModule)));
                 });
             }
 
@@ -201,9 +198,6 @@
                 return builder.ConfigureServices(services =>
                 {
                     services.AddApplicationInsightsTelemetry(o => o.RequestCollectionOptions.EnableW3CDistributedTracing = true);
-                    var depCollectorSd = services.Single(sd =>
-                        sd.ImplementationType == typeof(DependencyTrackingTelemetryModule));
-                    services.Remove(depCollectorSd);
                 });
             }))
             {
@@ -279,6 +273,12 @@
                     return builder.ConfigureServices(services =>
                     {
                         services.AddApplicationInsightsTelemetry(o => o.RequestCollectionOptions.EnableW3CDistributedTracing = true);
+                        services.ConfigureTelemetryModule<DependencyTrackingTelemetryModule>((m, o) =>
+                        {
+                            // no correlation headers so we can test request
+                            // call without auto-injected w3c headers
+                            m.ExcludeComponentCorrelationHttpHeadersOnDomains.Add("localhost");
+                        });
                     });
                 }))
             {